Fix DoubleSpinner rounding, add various comments
Fixed the issues I was facing with rounding with the DoubleSpinner (where the value wasn't being rounded even though it should be), by doing an extra round of rounding while stepping. I think that in most other circumstances (i.e. typing in a value), the rounding should automatically apply.

Also added various comments throughout the various spinner classes, so that I have more notes on how stuff gets done.

+++ SsuiAvalonia/LongSpinner.cs
@@ -237,22 +237,23 @@ namespace SolidShineUi
             _updateBox = false;
             if (DisplayAsHex)
             {
-                if (long.TryParse(txtValue.Text, System.Globalization.NumberStyles.HexNumber, null, out long newVal))
+                if (long.TryParse(txtValue.Text, System.Globalization.NumberStyles.HexNumber, null, out long newVal)) // the inputted value is a valid long, so let's make it the new value
                 {
                     Value = newVal;
                 }
             }
             else
             {
-                if (long.TryParse(txtValue.Text, System.Globalization.NumberStyles.Integer, null, out long newVal))
+                if (long.TryParse(txtValue.Text, System.Globalization.NumberStyles.Integer, null, out long newVal))  // the inputted value is a valid long, so let's make it the new value
                 {
                     Value = newVal;
+                    // Note: the property change system won't run again if the updated Value is actually the same as the current Value (i.e. the property didn't actually change)
                 }
-                else if (AcceptExpressions && ArithmeticParser.IsValidString(txtValue.Text))
+                else if (AcceptExpressions && ArithmeticParser.IsValidString(txtValue.Text)) // okay... not a number... maybe it's a math expression?
                 {
                     try
                     {
-                        Value = (int)Math.Round(ArithmeticParser.Evaluate(txtValue.Text), MidpointRounding.AwayFromZero);
+                        Value = (long)Math.Round(ArithmeticParser.Evaluate(txtValue.Text), MidpointRounding.AwayFromZero);
                     }
                     catch (FormatException)
                     {
